refactor: extract common-all git utils
