+++ packages/common-all/src/index.ts
@@ -39,6 +39,7 @@ export * from "./parse";
 export * from "./BacklinkUtils";
 export * from "./DLinkUtils";
 export * as YamlUtils from "./yaml";
+export * as GitUtils from "./git";
 
 export { axios, AxiosError };
 export { DateTime };
