+++ packages/common-server/src/git.ts
@@ -4,9 +4,8 @@ import {
   DVault,
   DWorkspace,
   NoteProps,
-  RESERVED_KEYS,
-  VaultUtils,
   ConfigUtils,
+  GitUtils as CommonGitUtils,
   FOLDERS,
   DendronConfig,
 } from "@dendronhq/common-all";
@@ -24,13 +23,6 @@ import { vault2Path } from "./filesv2";
 
 export { simpleGit, SimpleGit, SimpleGitResetMode };
 
-const formatString = (opts: { txt: string; note: NoteProps }) => {
-  const { txt, note } = opts;
-  _.templateSettings.interpolate = /{{([\s\S]+?)}}/g;
-  const noteHierarchy = note.fname.replace(/\./g, "/");
-  return _.template(txt)({ noteHierarchy });
-};
-
 /**
  *  NOTICE: Lots of the Git code is obtained from https://github.com/KnisterPeter/vscode-github, licened under MIT
  */
@@ -40,24 +32,7 @@ const formatString = (opts: { txt: string; note: NoteProps }) => {
  */
 export class GitUtils {
   static canShowGitLink(opts: { config: DendronConfig; note: NoteProps }) {
-    const { config, note } = opts;
-
-    if (
-      _.isBoolean((note.custom || {})[RESERVED_KEYS.GIT_NO_LINK]) &&
-      note.custom[RESERVED_KEYS.GIT_NO_LINK]
-    ) {
-      return false;
-    }
-    const githubConfig = ConfigUtils.getGithubConfig(config);
-    return githubConfig
-      ? _.every([
-          githubConfig.enableEditLink,
-          githubConfig.editLinkText,
-          githubConfig.editRepository,
-          githubConfig.editBranch,
-          githubConfig.editViewMode,
-        ])
-      : false;
+    return CommonGitUtils.canShowGitLink(opts);
   }
   /**
    * Convert a github repo orul to access token format
@@ -77,11 +52,7 @@ export class GitUtils {
   }
 
   static git2Github(gitUrl: string) {
-    // 'git@github.com:kevinslin/dendron-vault.git'
-    // @ts-ignore
-    const [_, userAndRepo] = gitUrl.split(":");
-    const [user, repo] = userAndRepo.split("/");
-    return `https://github.com/${user}/${path.basename(repo, ".git")}`;
+    return CommonGitUtils.git2Github(gitUrl);
   }
 
   static getGithubEditUrl(opts: {
@@ -89,46 +60,7 @@ export class GitUtils {
     config: DendronConfig;
     wsRoot: string;
   }) {
-    const { note, config, wsRoot } = opts;
-    const vault = note.vault;
-    const vaults = ConfigUtils.getVaults(config);
-    const mvault = VaultUtils.matchVault({ wsRoot, vault, vaults });
-    const vaultUrl = _.get(mvault, "remote.url", false);
-    const githubConfig = ConfigUtils.getGithubConfig(config);
-    const gitRepoUrl = githubConfig?.editRepository;
-    // if we have a vault, we don't need to include the vault name as an offset
-    if (mvault && vaultUrl) {
-      return _.join(
-        [
-          this.git2Github(vaultUrl),
-          githubConfig?.editViewMode,
-          githubConfig?.editBranch,
-          note.fname + ".md",
-        ],
-        "/"
-      );
-    }
-
-    let gitNotePath = _.join(
-      [path.basename(vault.fsPath), note.fname + ".md"],
-      "/"
-    );
-    if (_.has(note?.custom, RESERVED_KEYS.GIT_NOTE_PATH)) {
-      gitNotePath = formatString({
-        txt: note.custom[RESERVED_KEYS.GIT_NOTE_PATH],
-        note,
-      });
-    }
-    // this assumes we have a workspace url
-    return _.join(
-      [
-        gitRepoUrl,
-        githubConfig?.editViewMode,
-        githubConfig?.editBranch,
-        gitNotePath,
-      ],
-      "/"
-    );
+    return CommonGitUtils.getGithubEditUrl(opts);
   }
 
   static getOwnerAndRepoFromURL(url: string) {
